+++ CHANGELOG.md
@@ -6,7 +6,9 @@ CHANGELOG
 + new features
     - time triggered read tables implemented
 + polished
+    - some debug code removed at piDatabase.cpp
 + bugfixes
+    - some memory leaks fixed at piDatabase.cpp
 
 **v0.0.3: (2017/02/07)**
 + BETHA, app can now store INT tags values from MODBUS TCP/IP slave, and write to them.